pyproject.toml
[project]
name = "boax"
requires-python = ">=3.11"
description = "Boax is a Bayesian Optimization library for JAX."
keywords = []
authors = []
dependencies = [
    "numpy>=1.26.2",
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "chex>=0.1.84",
    "optax>=0.1.7",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version", "readme"]

[project.optional-dependencies]
docs = [
    "ipython>=8.18.1",
    "matplotlib>=3.8.2",
    "myst-nb>=1.0.0",
    "scikit-learn>=1.5.2",
    "sphinx>=6.2.1",
    "sphinx-autodoc-typehints>=1.23.0",
    "sphinx-book-theme>=1.1.2",
]

tests = [
    "absl-py>=2.0.0",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
]

[project.urls]
homepage = "https://github.com/Lando-L/boax"

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}
version = {attr = "boax.version.__version__"}

[tool.setuptools.packages.find]
include = ["boax*"]

[tool.ruff]
line-length = 80
indent-width = 2

[tool.ruff.lint]
select = ["F401", "I001"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
indent-style = "space"
quote-style = "single"